refactor: handle 파일 분리 및 수정
