+++ frontend/src/mocks/browser.js
@@ -1,6 +1,6 @@
 // src/mocks/browser.js
 import { setupWorker } from "msw";
-import { handlers } from "./handlers";
+import handlers from "./handlers/index";
 // This configures a Service Worker with the given request handlers.
 
 export const worker = setupWorker(...handlers);
