+++ frontend/src/mocks/handlers/index.js
@@ -1,0 +1,13 @@
+import memberHandlers from "./memberHandlers";
+import messageHandlers from "./messageHandlers";
+import rollingpaperHandlers from "./rollingpaperHandlers";
+import teamHandlers from "./teamHandlers";
+
+const handlers = [
+  ...memberHandlers,
+  ...messageHandlers,
+  ...rollingpaperHandlers,
+  ...teamHandlers,
+];
+
+export default handlers;
